backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
faiss-cpu
langchain
langchain-openai
